InProgress: Working on performance for studying.
-Most of bottleneck for GetNextStudyItem...() happens when using the
PhrasesByTextAndLanguageRetriever. This was slow because it was
using the PhraseList.GetAll() and then searching through those.
-I've added a method in the IPhraseDal (and propagated this throughout):
Fetch(text). This fetches all phrases that contain the text in any
language.
-I've also optimized (hopefully) the Fetch(ids), to where it does not
use FetchImpl(id) in a tight loop, but rather uses one sql (linq)
statement. I'm not sure if this works yet, cause I've not tested yet.
-Added a test to test for the Fetch(text) with the letter "l",
so it will try to fetch all phrases that contain the letter l in the test.
-Really tired, and my eye is twitching a little, so I'm leaving this off
for later.

Signed-off-by: William Raiford <bill.mybiz@gmail.com>