Implemented .ef and .mock userdal. changed languagedal stuff.
-Implemented IUserDal for .ef and .mock Dals.
-Added MustRunOnServerException.
-Added authorization check, DalHelper.CheckAuthorizationMustRunOnServer()
-Not sure if that is right, but I figure I am just trying to check
something when logging in. I can't check for authentication, because
these methods are called before authentication occurs. Will have to think
on this later.
-I've changed (or rather completed the change) to thinking of languages
as belonging to a single user. I was first contemplating having
shared languages for everyone, but I just don't like this. So, now, all
of my language dal stuff has checking for current user id, so no one
has shared resources on the db for anything now.
-Implemented the EfHelper methods for UserDto/Data.
-Added LanguageIds to UserDto. It was already in the UserData I believe.
-All tests are passing for the mock. I haven't checked for ef yet.

Signed-off-by: William Raiford <bill.mybiz@gmail.com>